Create the FileStorage and PathStorage classes that extend AbstractStorage.
